Move read ttl and ttlAction into CommonOptions
