+++ core/common/src/main/java/alluxio/thrift/GetStatusTOptions.java
@@ -40,8 +40,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
 
   private static final org.apache.thrift.protocol.TField LOAD_METADATA_TYPE_FIELD_DESC = new org.apache.thrift.protocol.TField("loadMetadataType", org.apache.thrift.protocol.TType.I32, (short)1);
   private static final org.apache.thrift.protocol.TField COMMON_OPTIONS_FIELD_DESC = new org.apache.thrift.protocol.TField("commonOptions", org.apache.thrift.protocol.TType.STRUCT, (short)2);
-  private static final org.apache.thrift.protocol.TField TTL_FIELD_DESC = new org.apache.thrift.protocol.TField("ttl", org.apache.thrift.protocol.TType.I64, (short)3);
-  private static final org.apache.thrift.protocol.TField TTL_ACTION_FIELD_DESC = new org.apache.thrift.protocol.TField("ttlAction", org.apache.thrift.protocol.TType.I32, (short)4);
 
   private static final Map<Class<? extends IScheme>, SchemeFactory> schemes = new HashMap<Class<? extends IScheme>, SchemeFactory>();
   static {
@@ -51,8 +49,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
 
   private LoadMetadataTType loadMetadataType; // optional
   private FileSystemMasterCommonTOptions commonOptions; // optional
-  private long ttl; // optional
-  private alluxio.thrift.TTtlAction ttlAction; // optional
 
   /** The set of fields this struct contains, along with convenience methods for finding and manipulating them. */
   public enum _Fields implements org.apache.thrift.TFieldIdEnum {
@@ -61,13 +57,7 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
      * @see LoadMetadataTType
      */
     LOAD_METADATA_TYPE((short)1, "loadMetadataType"),
-    COMMON_OPTIONS((short)2, "commonOptions"),
-    TTL((short)3, "ttl"),
-    /**
-     * 
-     * @see alluxio.thrift.TTtlAction
-     */
-    TTL_ACTION((short)4, "ttlAction");
+    COMMON_OPTIONS((short)2, "commonOptions");
 
     private static final Map<String, _Fields> byName = new HashMap<String, _Fields>();
 
@@ -86,10 +76,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
           return LOAD_METADATA_TYPE;
         case 2: // COMMON_OPTIONS
           return COMMON_OPTIONS;
-        case 3: // TTL
-          return TTL;
-        case 4: // TTL_ACTION
-          return TTL_ACTION;
         default:
           return null;
       }
@@ -130,9 +116,7 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
   }
 
   // isset id assignments
-  private static final int __TTL_ISSET_ID = 0;
-  private byte __isset_bitfield = 0;
-  private static final _Fields optionals[] = {_Fields.LOAD_METADATA_TYPE,_Fields.COMMON_OPTIONS,_Fields.TTL,_Fields.TTL_ACTION};
+  private static final _Fields optionals[] = {_Fields.LOAD_METADATA_TYPE,_Fields.COMMON_OPTIONS};
   public static final Map<_Fields, org.apache.thrift.meta_data.FieldMetaData> metaDataMap;
   static {
     Map<_Fields, org.apache.thrift.meta_data.FieldMetaData> tmpMap = new EnumMap<_Fields, org.apache.thrift.meta_data.FieldMetaData>(_Fields.class);
@@ -140,10 +124,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
         new org.apache.thrift.meta_data.EnumMetaData(org.apache.thrift.protocol.TType.ENUM, LoadMetadataTType.class)));
     tmpMap.put(_Fields.COMMON_OPTIONS, new org.apache.thrift.meta_data.FieldMetaData("commonOptions", org.apache.thrift.TFieldRequirementType.OPTIONAL, 
         new org.apache.thrift.meta_data.StructMetaData(org.apache.thrift.protocol.TType.STRUCT, FileSystemMasterCommonTOptions.class)));
-    tmpMap.put(_Fields.TTL, new org.apache.thrift.meta_data.FieldMetaData("ttl", org.apache.thrift.TFieldRequirementType.OPTIONAL, 
-        new org.apache.thrift.meta_data.FieldValueMetaData(org.apache.thrift.protocol.TType.I64)));
-    tmpMap.put(_Fields.TTL_ACTION, new org.apache.thrift.meta_data.FieldMetaData("ttlAction", org.apache.thrift.TFieldRequirementType.OPTIONAL, 
-        new org.apache.thrift.meta_data.EnumMetaData(org.apache.thrift.protocol.TType.ENUM, alluxio.thrift.TTtlAction.class)));
     metaDataMap = Collections.unmodifiableMap(tmpMap);
     org.apache.thrift.meta_data.FieldMetaData.addStructMetaDataMap(GetStatusTOptions.class, metaDataMap);
   }
@@ -155,16 +135,11 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
    * Performs a deep copy on <i>other</i>.
    */
   public GetStatusTOptions(GetStatusTOptions other) {
-    __isset_bitfield = other.__isset_bitfield;
     if (other.isSetLoadMetadataType()) {
       this.loadMetadataType = other.loadMetadataType;
     }
     if (other.isSetCommonOptions()) {
       this.commonOptions = new FileSystemMasterCommonTOptions(other.commonOptions);
-    }
-    this.ttl = other.ttl;
-    if (other.isSetTtlAction()) {
-      this.ttlAction = other.ttlAction;
     }
   }
 
@@ -176,9 +151,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
   public void clear() {
     this.loadMetadataType = null;
     this.commonOptions = null;
-    setTtlIsSet(false);
-    this.ttl = 0;
-    this.ttlAction = null;
   }
 
   /**
@@ -237,61 +209,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
     }
   }
 
-  public long getTtl() {
-    return this.ttl;
-  }
-
-  public GetStatusTOptions setTtl(long ttl) {
-    this.ttl = ttl;
-    setTtlIsSet(true);
-    return this;
-  }
-
-  public void unsetTtl() {
-    __isset_bitfield = EncodingUtils.clearBit(__isset_bitfield, __TTL_ISSET_ID);
-  }
-
-  /** Returns true if field ttl is set (has been assigned a value) and false otherwise */
-  public boolean isSetTtl() {
-    return EncodingUtils.testBit(__isset_bitfield, __TTL_ISSET_ID);
-  }
-
-  public void setTtlIsSet(boolean value) {
-    __isset_bitfield = EncodingUtils.setBit(__isset_bitfield, __TTL_ISSET_ID, value);
-  }
-
-  /**
-   * 
-   * @see alluxio.thrift.TTtlAction
-   */
-  public alluxio.thrift.TTtlAction getTtlAction() {
-    return this.ttlAction;
-  }
-
-  /**
-   * 
-   * @see alluxio.thrift.TTtlAction
-   */
-  public GetStatusTOptions setTtlAction(alluxio.thrift.TTtlAction ttlAction) {
-    this.ttlAction = ttlAction;
-    return this;
-  }
-
-  public void unsetTtlAction() {
-    this.ttlAction = null;
-  }
-
-  /** Returns true if field ttlAction is set (has been assigned a value) and false otherwise */
-  public boolean isSetTtlAction() {
-    return this.ttlAction != null;
-  }
-
-  public void setTtlActionIsSet(boolean value) {
-    if (!value) {
-      this.ttlAction = null;
-    }
-  }
-
   public void setFieldValue(_Fields field, Object value) {
     switch (field) {
     case LOAD_METADATA_TYPE:
@@ -310,22 +227,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
       }
       break;
 
-    case TTL:
-      if (value == null) {
-        unsetTtl();
-      } else {
-        setTtl((Long)value);
-      }
-      break;
-
-    case TTL_ACTION:
-      if (value == null) {
-        unsetTtlAction();
-      } else {
-        setTtlAction((alluxio.thrift.TTtlAction)value);
-      }
-      break;
-
     }
   }
 
@@ -336,12 +237,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
 
     case COMMON_OPTIONS:
       return getCommonOptions();
-
-    case TTL:
-      return getTtl();
-
-    case TTL_ACTION:
-      return getTtlAction();
 
     }
     throw new IllegalStateException();
@@ -358,10 +253,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
       return isSetLoadMetadataType();
     case COMMON_OPTIONS:
       return isSetCommonOptions();
-    case TTL:
-      return isSetTtl();
-    case TTL_ACTION:
-      return isSetTtlAction();
     }
     throw new IllegalStateException();
   }
@@ -397,24 +288,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
         return false;
     }
 
-    boolean this_present_ttl = true && this.isSetTtl();
-    boolean that_present_ttl = true && that.isSetTtl();
-    if (this_present_ttl || that_present_ttl) {
-      if (!(this_present_ttl && that_present_ttl))
-        return false;
-      if (this.ttl != that.ttl)
-        return false;
-    }
-
-    boolean this_present_ttlAction = true && this.isSetTtlAction();
-    boolean that_present_ttlAction = true && that.isSetTtlAction();
-    if (this_present_ttlAction || that_present_ttlAction) {
-      if (!(this_present_ttlAction && that_present_ttlAction))
-        return false;
-      if (!this.ttlAction.equals(that.ttlAction))
-        return false;
-    }
-
     return true;
   }
 
@@ -431,16 +304,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
     list.add(present_commonOptions);
     if (present_commonOptions)
       list.add(commonOptions);
-
-    boolean present_ttl = true && (isSetTtl());
-    list.add(present_ttl);
-    if (present_ttl)
-      list.add(ttl);
-
-    boolean present_ttlAction = true && (isSetTtlAction());
-    list.add(present_ttlAction);
-    if (present_ttlAction)
-      list.add(ttlAction.getValue());
 
     return list.hashCode();
   }
@@ -469,26 +332,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
     }
     if (isSetCommonOptions()) {
       lastComparison = org.apache.thrift.TBaseHelper.compareTo(this.commonOptions, other.commonOptions);
-      if (lastComparison != 0) {
-        return lastComparison;
-      }
-    }
-    lastComparison = Boolean.valueOf(isSetTtl()).compareTo(other.isSetTtl());
-    if (lastComparison != 0) {
-      return lastComparison;
-    }
-    if (isSetTtl()) {
-      lastComparison = org.apache.thrift.TBaseHelper.compareTo(this.ttl, other.ttl);
-      if (lastComparison != 0) {
-        return lastComparison;
-      }
-    }
-    lastComparison = Boolean.valueOf(isSetTtlAction()).compareTo(other.isSetTtlAction());
-    if (lastComparison != 0) {
-      return lastComparison;
-    }
-    if (isSetTtlAction()) {
-      lastComparison = org.apache.thrift.TBaseHelper.compareTo(this.ttlAction, other.ttlAction);
       if (lastComparison != 0) {
         return lastComparison;
       }
@@ -532,22 +375,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
       }
       first = false;
     }
-    if (isSetTtl()) {
-      if (!first) sb.append(", ");
-      sb.append("ttl:");
-      sb.append(this.ttl);
-      first = false;
-    }
-    if (isSetTtlAction()) {
-      if (!first) sb.append(", ");
-      sb.append("ttlAction:");
-      if (this.ttlAction == null) {
-        sb.append("null");
-      } else {
-        sb.append(this.ttlAction);
-      }
-      first = false;
-    }
     sb.append(")");
     return sb.toString();
   }
@@ -570,8 +397,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
 
   private void readObject(java.io.ObjectInputStream in) throws java.io.IOException, ClassNotFoundException {
     try {
-      // it doesn't seem like you should have to do this, but java serialization is wacky, and doesn't call the default constructor.
-      __isset_bitfield = 0;
       read(new org.apache.thrift.protocol.TCompactProtocol(new org.apache.thrift.transport.TIOStreamTransport(in)));
     } catch (org.apache.thrift.TException te) {
       throw new java.io.IOException(te);
@@ -613,22 +438,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
               org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
             }
             break;
-          case 3: // TTL
-            if (schemeField.type == org.apache.thrift.protocol.TType.I64) {
-              struct.ttl = iprot.readI64();
-              struct.setTtlIsSet(true);
-            } else { 
-              org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
-            }
-            break;
-          case 4: // TTL_ACTION
-            if (schemeField.type == org.apache.thrift.protocol.TType.I32) {
-              struct.ttlAction = alluxio.thrift.TTtlAction.findByValue(iprot.readI32());
-              struct.setTtlActionIsSet(true);
-            } else { 
-              org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
-            }
-            break;
           default:
             org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
         }
@@ -658,18 +467,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
           oprot.writeFieldEnd();
         }
       }
-      if (struct.isSetTtl()) {
-        oprot.writeFieldBegin(TTL_FIELD_DESC);
-        oprot.writeI64(struct.ttl);
-        oprot.writeFieldEnd();
-      }
-      if (struct.ttlAction != null) {
-        if (struct.isSetTtlAction()) {
-          oprot.writeFieldBegin(TTL_ACTION_FIELD_DESC);
-          oprot.writeI32(struct.ttlAction.getValue());
-          oprot.writeFieldEnd();
-        }
-      }
       oprot.writeFieldStop();
       oprot.writeStructEnd();
     }
@@ -694,31 +491,19 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
       if (struct.isSetCommonOptions()) {
         optionals.set(1);
       }
-      if (struct.isSetTtl()) {
-        optionals.set(2);
-      }
-      if (struct.isSetTtlAction()) {
-        optionals.set(3);
-      }
-      oprot.writeBitSet(optionals, 4);
+      oprot.writeBitSet(optionals, 2);
       if (struct.isSetLoadMetadataType()) {
         oprot.writeI32(struct.loadMetadataType.getValue());
       }
       if (struct.isSetCommonOptions()) {
         struct.commonOptions.write(oprot);
       }
-      if (struct.isSetTtl()) {
-        oprot.writeI64(struct.ttl);
-      }
-      if (struct.isSetTtlAction()) {
-        oprot.writeI32(struct.ttlAction.getValue());
-      }
     }
 
     @Override
     public void read(org.apache.thrift.protocol.TProtocol prot, GetStatusTOptions struct) throws org.apache.thrift.TException {
       TTupleProtocol iprot = (TTupleProtocol) prot;
-      BitSet incoming = iprot.readBitSet(4);
+      BitSet incoming = iprot.readBitSet(2);
       if (incoming.get(0)) {
         struct.loadMetadataType = alluxio.thrift.LoadMetadataTType.findByValue(iprot.readI32());
         struct.setLoadMetadataTypeIsSet(true);
@@ -727,14 +512,6 @@ public class GetStatusTOptions implements org.apache.thrift.TBase<GetStatusTOpti
         struct.commonOptions = new FileSystemMasterCommonTOptions();
         struct.commonOptions.read(iprot);
         struct.setCommonOptionsIsSet(true);
-      }
-      if (incoming.get(2)) {
-        struct.ttl = iprot.readI64();
-        struct.setTtlIsSet(true);
-      }
-      if (incoming.get(3)) {
-        struct.ttlAction = alluxio.thrift.TTtlAction.findByValue(iprot.readI32());
-        struct.setTtlActionIsSet(true);
       }
     }
   }
