+++ core/common/src/main/java/alluxio/thrift/ListStatusTOptions.java
@@ -41,8 +41,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
   private static final org.apache.thrift.protocol.TField LOAD_DIRECT_CHILDREN_FIELD_DESC = new org.apache.thrift.protocol.TField("loadDirectChildren", org.apache.thrift.protocol.TType.BOOL, (short)1);
   private static final org.apache.thrift.protocol.TField LOAD_METADATA_TYPE_FIELD_DESC = new org.apache.thrift.protocol.TField("loadMetadataType", org.apache.thrift.protocol.TType.I32, (short)2);
   private static final org.apache.thrift.protocol.TField COMMON_OPTIONS_FIELD_DESC = new org.apache.thrift.protocol.TField("commonOptions", org.apache.thrift.protocol.TType.STRUCT, (short)3);
-  private static final org.apache.thrift.protocol.TField TTL_FIELD_DESC = new org.apache.thrift.protocol.TField("ttl", org.apache.thrift.protocol.TType.I64, (short)4);
-  private static final org.apache.thrift.protocol.TField TTL_ACTION_FIELD_DESC = new org.apache.thrift.protocol.TField("ttlAction", org.apache.thrift.protocol.TType.I32, (short)5);
 
   private static final Map<Class<? extends IScheme>, SchemeFactory> schemes = new HashMap<Class<? extends IScheme>, SchemeFactory>();
   static {
@@ -53,8 +51,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
   private boolean loadDirectChildren; // optional
   private LoadMetadataTType loadMetadataType; // optional
   private FileSystemMasterCommonTOptions commonOptions; // optional
-  private long ttl; // optional
-  private alluxio.thrift.TTtlAction ttlAction; // optional
 
   /** The set of fields this struct contains, along with convenience methods for finding and manipulating them. */
   public enum _Fields implements org.apache.thrift.TFieldIdEnum {
@@ -64,13 +60,7 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
      * @see LoadMetadataTType
      */
     LOAD_METADATA_TYPE((short)2, "loadMetadataType"),
-    COMMON_OPTIONS((short)3, "commonOptions"),
-    TTL((short)4, "ttl"),
-    /**
-     * 
-     * @see alluxio.thrift.TTtlAction
-     */
-    TTL_ACTION((short)5, "ttlAction");
+    COMMON_OPTIONS((short)3, "commonOptions");
 
     private static final Map<String, _Fields> byName = new HashMap<String, _Fields>();
 
@@ -91,10 +81,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
           return LOAD_METADATA_TYPE;
         case 3: // COMMON_OPTIONS
           return COMMON_OPTIONS;
-        case 4: // TTL
-          return TTL;
-        case 5: // TTL_ACTION
-          return TTL_ACTION;
         default:
           return null;
       }
@@ -136,9 +122,8 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
 
   // isset id assignments
   private static final int __LOADDIRECTCHILDREN_ISSET_ID = 0;
-  private static final int __TTL_ISSET_ID = 1;
   private byte __isset_bitfield = 0;
-  private static final _Fields optionals[] = {_Fields.LOAD_DIRECT_CHILDREN,_Fields.LOAD_METADATA_TYPE,_Fields.COMMON_OPTIONS,_Fields.TTL,_Fields.TTL_ACTION};
+  private static final _Fields optionals[] = {_Fields.LOAD_DIRECT_CHILDREN,_Fields.LOAD_METADATA_TYPE,_Fields.COMMON_OPTIONS};
   public static final Map<_Fields, org.apache.thrift.meta_data.FieldMetaData> metaDataMap;
   static {
     Map<_Fields, org.apache.thrift.meta_data.FieldMetaData> tmpMap = new EnumMap<_Fields, org.apache.thrift.meta_data.FieldMetaData>(_Fields.class);
@@ -148,10 +133,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
         new org.apache.thrift.meta_data.EnumMetaData(org.apache.thrift.protocol.TType.ENUM, LoadMetadataTType.class)));
     tmpMap.put(_Fields.COMMON_OPTIONS, new org.apache.thrift.meta_data.FieldMetaData("commonOptions", org.apache.thrift.TFieldRequirementType.OPTIONAL, 
         new org.apache.thrift.meta_data.StructMetaData(org.apache.thrift.protocol.TType.STRUCT, FileSystemMasterCommonTOptions.class)));
-    tmpMap.put(_Fields.TTL, new org.apache.thrift.meta_data.FieldMetaData("ttl", org.apache.thrift.TFieldRequirementType.OPTIONAL, 
-        new org.apache.thrift.meta_data.FieldValueMetaData(org.apache.thrift.protocol.TType.I64)));
-    tmpMap.put(_Fields.TTL_ACTION, new org.apache.thrift.meta_data.FieldMetaData("ttlAction", org.apache.thrift.TFieldRequirementType.OPTIONAL, 
-        new org.apache.thrift.meta_data.EnumMetaData(org.apache.thrift.protocol.TType.ENUM, alluxio.thrift.TTtlAction.class)));
     metaDataMap = Collections.unmodifiableMap(tmpMap);
     org.apache.thrift.meta_data.FieldMetaData.addStructMetaDataMap(ListStatusTOptions.class, metaDataMap);
   }
@@ -171,10 +152,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
     if (other.isSetCommonOptions()) {
       this.commonOptions = new FileSystemMasterCommonTOptions(other.commonOptions);
     }
-    this.ttl = other.ttl;
-    if (other.isSetTtlAction()) {
-      this.ttlAction = other.ttlAction;
-    }
   }
 
   public ListStatusTOptions deepCopy() {
@@ -187,9 +164,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
     this.loadDirectChildren = false;
     this.loadMetadataType = null;
     this.commonOptions = null;
-    setTtlIsSet(false);
-    this.ttl = 0;
-    this.ttlAction = null;
   }
 
   public boolean isLoadDirectChildren() {
@@ -271,61 +245,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
     }
   }
 
-  public long getTtl() {
-    return this.ttl;
-  }
-
-  public ListStatusTOptions setTtl(long ttl) {
-    this.ttl = ttl;
-    setTtlIsSet(true);
-    return this;
-  }
-
-  public void unsetTtl() {
-    __isset_bitfield = EncodingUtils.clearBit(__isset_bitfield, __TTL_ISSET_ID);
-  }
-
-  /** Returns true if field ttl is set (has been assigned a value) and false otherwise */
-  public boolean isSetTtl() {
-    return EncodingUtils.testBit(__isset_bitfield, __TTL_ISSET_ID);
-  }
-
-  public void setTtlIsSet(boolean value) {
-    __isset_bitfield = EncodingUtils.setBit(__isset_bitfield, __TTL_ISSET_ID, value);
-  }
-
-  /**
-   * 
-   * @see alluxio.thrift.TTtlAction
-   */
-  public alluxio.thrift.TTtlAction getTtlAction() {
-    return this.ttlAction;
-  }
-
-  /**
-   * 
-   * @see alluxio.thrift.TTtlAction
-   */
-  public ListStatusTOptions setTtlAction(alluxio.thrift.TTtlAction ttlAction) {
-    this.ttlAction = ttlAction;
-    return this;
-  }
-
-  public void unsetTtlAction() {
-    this.ttlAction = null;
-  }
-
-  /** Returns true if field ttlAction is set (has been assigned a value) and false otherwise */
-  public boolean isSetTtlAction() {
-    return this.ttlAction != null;
-  }
-
-  public void setTtlActionIsSet(boolean value) {
-    if (!value) {
-      this.ttlAction = null;
-    }
-  }
-
   public void setFieldValue(_Fields field, Object value) {
     switch (field) {
     case LOAD_DIRECT_CHILDREN:
@@ -352,22 +271,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
       }
       break;
 
-    case TTL:
-      if (value == null) {
-        unsetTtl();
-      } else {
-        setTtl((Long)value);
-      }
-      break;
-
-    case TTL_ACTION:
-      if (value == null) {
-        unsetTtlAction();
-      } else {
-        setTtlAction((alluxio.thrift.TTtlAction)value);
-      }
-      break;
-
     }
   }
 
@@ -381,12 +284,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
 
     case COMMON_OPTIONS:
       return getCommonOptions();
-
-    case TTL:
-      return getTtl();
-
-    case TTL_ACTION:
-      return getTtlAction();
 
     }
     throw new IllegalStateException();
@@ -405,10 +302,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
       return isSetLoadMetadataType();
     case COMMON_OPTIONS:
       return isSetCommonOptions();
-    case TTL:
-      return isSetTtl();
-    case TTL_ACTION:
-      return isSetTtlAction();
     }
     throw new IllegalStateException();
   }
@@ -453,24 +346,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
         return false;
     }
 
-    boolean this_present_ttl = true && this.isSetTtl();
-    boolean that_present_ttl = true && that.isSetTtl();
-    if (this_present_ttl || that_present_ttl) {
-      if (!(this_present_ttl && that_present_ttl))
-        return false;
-      if (this.ttl != that.ttl)
-        return false;
-    }
-
-    boolean this_present_ttlAction = true && this.isSetTtlAction();
-    boolean that_present_ttlAction = true && that.isSetTtlAction();
-    if (this_present_ttlAction || that_present_ttlAction) {
-      if (!(this_present_ttlAction && that_present_ttlAction))
-        return false;
-      if (!this.ttlAction.equals(that.ttlAction))
-        return false;
-    }
-
     return true;
   }
 
@@ -492,16 +367,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
     list.add(present_commonOptions);
     if (present_commonOptions)
       list.add(commonOptions);
-
-    boolean present_ttl = true && (isSetTtl());
-    list.add(present_ttl);
-    if (present_ttl)
-      list.add(ttl);
-
-    boolean present_ttlAction = true && (isSetTtlAction());
-    list.add(present_ttlAction);
-    if (present_ttlAction)
-      list.add(ttlAction.getValue());
 
     return list.hashCode();
   }
@@ -540,26 +405,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
     }
     if (isSetCommonOptions()) {
       lastComparison = org.apache.thrift.TBaseHelper.compareTo(this.commonOptions, other.commonOptions);
-      if (lastComparison != 0) {
-        return lastComparison;
-      }
-    }
-    lastComparison = Boolean.valueOf(isSetTtl()).compareTo(other.isSetTtl());
-    if (lastComparison != 0) {
-      return lastComparison;
-    }
-    if (isSetTtl()) {
-      lastComparison = org.apache.thrift.TBaseHelper.compareTo(this.ttl, other.ttl);
-      if (lastComparison != 0) {
-        return lastComparison;
-      }
-    }
-    lastComparison = Boolean.valueOf(isSetTtlAction()).compareTo(other.isSetTtlAction());
-    if (lastComparison != 0) {
-      return lastComparison;
-    }
-    if (isSetTtlAction()) {
-      lastComparison = org.apache.thrift.TBaseHelper.compareTo(this.ttlAction, other.ttlAction);
       if (lastComparison != 0) {
         return lastComparison;
       }
@@ -606,22 +451,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
         sb.append("null");
       } else {
         sb.append(this.commonOptions);
-      }
-      first = false;
-    }
-    if (isSetTtl()) {
-      if (!first) sb.append(", ");
-      sb.append("ttl:");
-      sb.append(this.ttl);
-      first = false;
-    }
-    if (isSetTtlAction()) {
-      if (!first) sb.append(", ");
-      sb.append("ttlAction:");
-      if (this.ttlAction == null) {
-        sb.append("null");
-      } else {
-        sb.append(this.ttlAction);
       }
       first = false;
     }
@@ -698,22 +527,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
               org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
             }
             break;
-          case 4: // TTL
-            if (schemeField.type == org.apache.thrift.protocol.TType.I64) {
-              struct.ttl = iprot.readI64();
-              struct.setTtlIsSet(true);
-            } else { 
-              org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
-            }
-            break;
-          case 5: // TTL_ACTION
-            if (schemeField.type == org.apache.thrift.protocol.TType.I32) {
-              struct.ttlAction = alluxio.thrift.TTtlAction.findByValue(iprot.readI32());
-              struct.setTtlActionIsSet(true);
-            } else { 
-              org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
-            }
-            break;
           default:
             org.apache.thrift.protocol.TProtocolUtil.skip(iprot, schemeField.type);
         }
@@ -748,18 +561,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
           oprot.writeFieldEnd();
         }
       }
-      if (struct.isSetTtl()) {
-        oprot.writeFieldBegin(TTL_FIELD_DESC);
-        oprot.writeI64(struct.ttl);
-        oprot.writeFieldEnd();
-      }
-      if (struct.ttlAction != null) {
-        if (struct.isSetTtlAction()) {
-          oprot.writeFieldBegin(TTL_ACTION_FIELD_DESC);
-          oprot.writeI32(struct.ttlAction.getValue());
-          oprot.writeFieldEnd();
-        }
-      }
       oprot.writeFieldStop();
       oprot.writeStructEnd();
     }
@@ -787,13 +588,7 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
       if (struct.isSetCommonOptions()) {
         optionals.set(2);
       }
-      if (struct.isSetTtl()) {
-        optionals.set(3);
-      }
-      if (struct.isSetTtlAction()) {
-        optionals.set(4);
-      }
-      oprot.writeBitSet(optionals, 5);
+      oprot.writeBitSet(optionals, 3);
       if (struct.isSetLoadDirectChildren()) {
         oprot.writeBool(struct.loadDirectChildren);
       }
@@ -803,18 +598,12 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
       if (struct.isSetCommonOptions()) {
         struct.commonOptions.write(oprot);
       }
-      if (struct.isSetTtl()) {
-        oprot.writeI64(struct.ttl);
-      }
-      if (struct.isSetTtlAction()) {
-        oprot.writeI32(struct.ttlAction.getValue());
-      }
     }
 
     @Override
     public void read(org.apache.thrift.protocol.TProtocol prot, ListStatusTOptions struct) throws org.apache.thrift.TException {
       TTupleProtocol iprot = (TTupleProtocol) prot;
-      BitSet incoming = iprot.readBitSet(5);
+      BitSet incoming = iprot.readBitSet(3);
       if (incoming.get(0)) {
         struct.loadDirectChildren = iprot.readBool();
         struct.setLoadDirectChildrenIsSet(true);
@@ -827,14 +616,6 @@ public class ListStatusTOptions implements org.apache.thrift.TBase<ListStatusTOp
         struct.commonOptions = new FileSystemMasterCommonTOptions();
         struct.commonOptions.read(iprot);
         struct.setCommonOptionsIsSet(true);
-      }
-      if (incoming.get(3)) {
-        struct.ttl = iprot.readI64();
-        struct.setTtlIsSet(true);
-      }
-      if (incoming.get(4)) {
-        struct.ttlAction = alluxio.thrift.TTtlAction.findByValue(iprot.readI32());
-        struct.setTtlActionIsSet(true);
       }
     }
   }
