+++ core/common/src/thrift/file_system_master.thrift
@@ -5,6 +5,8 @@ include "exception.thrift"
 
 struct FileSystemMasterCommonTOptions {
   1: optional i64 syncIntervalMs
+  2: optional i64 ttl
+  3: optional common.TTtlAction ttlAction
 }
 
 struct CheckConsistencyTOptions {
@@ -73,8 +75,6 @@ struct GetNewBlockIdForFileTResponse {
 struct GetStatusTOptions {
   1: optional LoadMetadataTType loadMetadataType
   2: optional FileSystemMasterCommonTOptions commonOptions
-  3: optional i64 ttl
-  4: optional common.TTtlAction ttlAction
 }
 struct GetStatusTResponse {
   1: FileInfo fileInfo
@@ -85,8 +85,6 @@ struct ListStatusTOptions {
   1: optional bool loadDirectChildren
   2: optional LoadMetadataTType loadMetadataType
   3: optional FileSystemMasterCommonTOptions commonOptions
-  4: optional i64 ttl
-  5: optional common.TTtlAction ttlAction
 }
 struct ListStatusTResponse {
   1: list<FileInfo> fileInfoList
