+++ core/server/master/src/main/java/alluxio/master/file/options/ListStatusOptions.java
@@ -11,11 +11,9 @@
 
 package alluxio.master.file.options;
 
-import alluxio.Constants;
 import alluxio.thrift.ListStatusTOptions;
 import alluxio.wire.CommonOptions;
 import alluxio.wire.LoadMetadataType;
-import alluxio.wire.TtlAction;
 
 import com.google.common.base.Objects;
 
@@ -28,8 +26,6 @@ import javax.annotation.concurrent.NotThreadSafe;
 public final class ListStatusOptions {
   private CommonOptions mCommonOptions;
   private LoadMetadataType mLoadMetadataType;
-  private long mTtl;
-  private TtlAction mTtlAction;
 
   /**
    * @return the default {@link ListStatusOptions}
@@ -42,8 +38,6 @@ public final class ListStatusOptions {
     super();
     mCommonOptions = CommonOptions.defaults();
     mLoadMetadataType = LoadMetadataType.Once;
-    mTtl = Constants.NO_TTL;
-    mTtlAction = TtlAction.DELETE;
   }
 
   /**
@@ -62,8 +56,6 @@ public final class ListStatusOptions {
       } else if (!options.isLoadDirectChildren()) {
         mLoadMetadataType = LoadMetadataType.Never;
       }
-      mTtl = options.getTtl();
-      mTtlAction = TtlAction.fromThrift(options.getTtlAction());
     }
   }
 
@@ -80,20 +72,6 @@ public final class ListStatusOptions {
    */
   public LoadMetadataType getLoadMetadataType() {
     return mLoadMetadataType;
-  }
-
-  /**
-   * @return time to live
-   */
-  public long getTtl() {
-    return mTtl;
-  }
-
-  /**
-   * @return action after ttl expired
-   */
-  public TtlAction getTtlAction() {
-    return mTtlAction;
   }
 
   /**
@@ -116,20 +94,6 @@ public final class ListStatusOptions {
     return this;
   }
 
-  /**
-   * @param ttl time to live
-   */
-  public void setTtl(long ttl) {
-    mTtl = ttl;
-  }
-
-  /**
-   * @param ttlAction action after ttl expired
-   */
-  public void setTtlAction(TtlAction ttlAction) {
-    mTtlAction = ttlAction;
-  }
-
   @Override
   public boolean equals(Object o) {
     if (this == o) {
@@ -140,14 +104,12 @@ public final class ListStatusOptions {
     }
     ListStatusOptions that = (ListStatusOptions) o;
     return Objects.equal(mLoadMetadataType, that.mLoadMetadataType)
-        && Objects.equal(mCommonOptions, that.mCommonOptions)
-        && Objects.equal(mTtl, that.mTtl)
-        && Objects.equal(mTtlAction, that.mTtlAction);
+        && Objects.equal(mCommonOptions, that.mCommonOptions);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hashCode(mLoadMetadataType, mCommonOptions, mTtl, mTtlAction);
+    return Objects.hashCode(mLoadMetadataType, mCommonOptions);
   }
 
   @Override
@@ -155,8 +117,6 @@ public final class ListStatusOptions {
     return Objects.toStringHelper(this)
         .add("commonOptions", mCommonOptions)
         .add("loadMetadataType", mLoadMetadataType.toString())
-        .add("ttl", mTtl)
-        .add("ttlAction", mTtlAction.toString())
         .toString();
   }
 }
